Extract Utility Classes to Test.Utility
