+++ test/NuGet.Clients.Tests/NuGet.PackageManagement.VisualStudio.Test/ProjectSystems/LegacyPackageReferenceProjectTests.cs
@@ -24,7 +24,7 @@ using NuGet.ProjectModel;
 using NuGet.Test.Utility;
 using NuGet.Versioning;
 using NuGet.VisualStudio;
-using Test.Utility.VisualStudio;
+using Test.Utility;
 using Xunit;
 using static NuGet.PackageManagement.VisualStudio.Test.ProjectFactories;
 using Task = System.Threading.Tasks.Task;
@@ -1590,62 +1590,5 @@ namespace NuGet.PackageManagement.VisualStudio.Test
                 _threadingService);
             return testProject;
         }
-    }
-
-    internal class TestProjectSystemServices : INuGetProjectServices
-    {
-        public TestProjectSystemServices()
-        {
-            Mock.Get(ReferencesReader)
-                .Setup(x => x.GetProjectReferencesAsync(
-                    It.IsAny<NuGet.Common.ILogger>(), CancellationToken.None))
-                .ReturnsAsync(() => new ProjectRestoreReference[] { });
-
-            Mock.Get(ReferencesReader)
-                .Setup(x => x.GetPackageReferencesAsync(
-                    It.IsAny<NuGetFramework>(), CancellationToken.None))
-                .ReturnsAsync(() => new LibraryDependency[] { });
-        }
-
-        public IProjectBuildProperties BuildProperties { get; } = Mock.Of<IProjectBuildProperties>();
-
-        public IProjectSystemCapabilities Capabilities { get; } = Mock.Of<IProjectSystemCapabilities>();
-
-        public IProjectSystemReferencesReader ReferencesReader { get; } = Mock.Of<IProjectSystemReferencesReader>();
-
-        public IProjectSystemService ProjectSystem { get; } = Mock.Of<IProjectSystemService>();
-
-        public IProjectSystemReferencesService References { get; } = Mock.Of<IProjectSystemReferencesService>();
-
-        public IProjectScriptHostService ScriptService { get; } = Mock.Of<IProjectScriptHostService>();
-
-        public T GetGlobalService<T>() where T : class
-        {
-            throw new NotImplementedException();
-        }
-
-        public void SetupInstalledPackages(NuGetFramework targetFramework, params LibraryDependency[] dependencies)
-        {
-            Mock.Get(ReferencesReader)
-                .Setup(x => x.GetPackageReferencesAsync(targetFramework, CancellationToken.None))
-                .ReturnsAsync(dependencies.ToList());
-        }
-
-        public void SetupProjectDependencies(params ProjectRestoreReference[] dependencies)
-        {
-            Mock.Get(ReferencesReader)
-                .Setup(x => x.GetProjectReferencesAsync(It.IsAny<NuGet.Common.ILogger>(), CancellationToken.None))
-                .ReturnsAsync(dependencies.ToList());
-        }
-    }
-
-    public class TestProjectThreadingService : IVsProjectThreadingService
-    {
-        public TestProjectThreadingService(JoinableTaskFactory jtf)
-        {
-            JoinableTaskFactory = jtf;
-        }
-
-        public JoinableTaskFactory JoinableTaskFactory { get; }
     }
 }
