+++ test/TestUtilities/Test.Utility/ProjectManagement/TestProjectThreadingService.cs
@@ -1,0 +1,18 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using Microsoft.VisualStudio.Threading;
+using NuGet.PackageManagement.VisualStudio;
+
+namespace Test.Utility
+{
+    public class TestProjectThreadingService : IVsProjectThreadingService
+    {
+        public TestProjectThreadingService(JoinableTaskFactory jtf)
+        {
+            JoinableTaskFactory = jtf;
+        }
+
+        public JoinableTaskFactory JoinableTaskFactory { get; }
+    }
+}
